+++ src/Component/Provider/MicrosoftAzureClient.php
@@ -3,7 +3,7 @@
 namespace Heptacom\AdminOpenAuth\Component\Provider;
 
 use Heptacom\AdminOpenAuth\Component\Contract\ClientInterface;
-use Heptacom\AdminOpenAuth\Contract\TokenPairFactoryInterface;
+use Heptacom\OpenAuth\Client\Contract\TokenPairFactoryContract;
 use Heptacom\OpenAuth\Struct\TokenPairStruct;
 use Heptacom\OpenAuth\Struct\UserStruct;
 use League\OAuth2\Client\Provider\AbstractProvider;
@@ -12,7 +12,7 @@ use TheNetworg\OAuth2\Client\Provider\Azure;
 class MicrosoftAzureClient implements ClientInterface
 {
     /**
-     * @var TokenPairFactoryInterface
+     * @var TokenPairFactoryContract
      */
     private $tokenPairFactory;
 
@@ -21,7 +21,7 @@ class MicrosoftAzureClient implements ClientInterface
      */
     private $azureClient;
 
-    public function __construct(TokenPairFactoryInterface $tokenPairFactory, array $options)
+    public function __construct(TokenPairFactoryContract $tokenPairFactory, array $options)
     {
         $this->tokenPairFactory = $tokenPairFactory;
         $this->azureClient = new Azure($options);